mcp-shenanigans/requirements.txt
httpx==0.26.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4 